Refact: split swagger config
