+++ src/config/swagger-config.ts
@@ -1,0 +1,13 @@
+import { INestApplication } from "@nestjs/common";
+import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger";
+
+export function initSwaggerConfig(app: INestApplication): void {
+    const config = new DocumentBuilder()
+    .setTitle('title example')
+    .setDescription('description example')
+    .setVersion('1.0.0')
+    .build();
+  
+    const document = SwaggerModule.createDocument(app, config);
+    SwaggerModule.setup('api-docs', app, document);
+}
