+++ src/main.ts
@@ -1,9 +1,9 @@
-import { ValidationPipe, INestApplication } from '@nestjs/common';
+import { ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { WinstonModule } from 'nest-winston';
 import { format, transports } from 'winston';
+import { initSwaggerConfig } from './config/swagger-config';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule, {
@@ -43,17 +43,6 @@ async function bootstrap() {
   initSwaggerConfig(app);
 
   await app.listen(3000);
-}
-
-function initSwaggerConfig(app: INestApplication): void {
-  const config = new DocumentBuilder()
-  .setTitle('title example')
-  .setDescription('description example')
-  .setVersion('1.0.0')
-  .build();
-
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api-docs', app, document);
 }
 
 bootstrap();
